Changes for parent-child relationship

+++ routes/api/category.js
@@ -3,6 +3,7 @@ const router = express.Router();
 const auth = require('../../middleware/auth');
 const { validationResult } = require('express-validator');
 const Category = require('../../models/Category');
+const SubCategory = require('../../models/SubCategory');
 
 // @route   POST api/category
 // @desc    Create a category
@@ -116,11 +117,16 @@ router.post('/:category_id', auth, async (req, res) => {
 
 router.delete('/:category_id', auth, async (req, res) => {
   try {
+    const subcategory = await SubCategory.find({categoryId:req.params.category_id})
+    if(subcategory){
+      return res.json({msg:'Category has sub categories available and hence cannot be deleted!'})
+    }
     await Category.findOneAndRemove({
       _id: req.params.category_id
     });
     res.json({ msg: 'Category deleted' });
   } catch (err) {
+    console.log('delete error',err);
     /* istanbul ignore next */
     res.status(500).send('Server Error');
   }
